cleanup for SearchHeader component
